Reviewed-on: https://gitea.vodsystem.ddns.net/dylan/vod-system/pulls/8

+++ src/main/java/com/ddf/vodsystem/dto/ClipOptions.java
@@ -27,5 +27,5 @@ public class ClipOptions {
     private Integer height;
 
     @Positive
-    private Float fileSize;
+    private Long fileSize;
 }
